feat: add steps to revert/delete generated vhost if apache config check errors
if it works as intended I'd make an own taskbook out of it that accepts vhost dest as an "input" (var:)

this would be so much simpler and more correct if there was a comment to just check a single file.
this solution is bad in the way that it fails even when the error is outside of its control,
e.g. someone messed up apache config somewhere else
but this solution stops from reloading apache if there is an error,
which is way better